print_tf_vars/requirements.txt
pyhcl==0.3.9

